Merge from upstream through 2020-09-15

+++ ReactCommon/react/renderer/textlayoutmanager/platform/android/react/renderer/textlayoutmanager/TextLayoutManager.h
@@ -42,7 +42,7 @@ class TextLayoutManager {
    * opaque cache ID.
    */
   TextMeasurement measureCachedSpannableById(
-      int cacheId,
+      int64_t cacheId,
       ParagraphAttributes paragraphAttributes,
       LayoutConstraints layoutConstraints) const;
